requirements.txt
# Core dependencies
pandas>=2.2.3,<3.0.0
numpy>=2.1.2,<3.0.0
scikit-learn>=1.5.2,<2.0.0
matplotlib>=3.9.2,<4.0.0
seaborn>=0.13.2,<1.0.0
streamlit>=1.39.0,<2.0.0
joblib>=1.3.0,<2.0.0

# Data handling and configuration
pyyaml>=6.0.1,<7.0.0

# Optional: for improved performance
numba>=0.58.0,<1.0.0  # For accelerated computations
pandas-datareader>=0.10.0  # For financial data handling

# Testing
pytest>=7.4.0,<8.0.0
pytest-cov>=4.1.0,<5.0.0  # For test coverage reports

# Development
black>=23.3.0,<24.0.0  # Code formatting
isort>=5.12.0,<6.0.0  # Import sorting
flake8>=6.0.0,<7.0.0  # Style checking
mypy>=1.5.1,<2.0.0  # Type checking
pre-commit>=3.3.3,<4.0.0  # Pre-commit hooks

# Environment and config
python-dotenv>=1.0.0,<2.0.0
